(#1146) WIP moved populate_ free functions into callbacks layer
