+++ src/hyperion/external_interaction/ispyb/rotation_ispyb_store.py
@@ -4,14 +4,12 @@ import ispyb
 
 from hyperion.external_interaction.ispyb.data_model import (
     DataCollectionGroupInfo,
-    DataCollectionInfo,
     ScanDataInfo,
 )
 from hyperion.external_interaction.ispyb.ispyb_store import (
     IspybIds,
     StoreInIspyb,
 )
-from hyperion.external_interaction.ispyb.ispyb_utils import get_xtal_snapshots
 
 
 class StoreRotationInIspyb(StoreInIspyb):
@@ -79,34 +77,8 @@ class StoreRotationInIspyb(StoreInIspyb):
                 data_collection_ids=(ids[0],), data_collection_group_id=ids[1]
             )
 
-    def end_deposition(self, success: str, reason: str, internal_params):
+    def end_deposition(self, success: str, reason: str):
         assert (
             self._data_collection_id is not None
         ), "Can't end ISPyB deposition, data_collection IDs is missing"
         self._end_deposition(self._data_collection_id, success, reason)
-
-
-def populate_data_collection_info_for_rotation(
-    ispyb_params, detector_params, full_params
-):
-    info = DataCollectionInfo(
-        omega_start=detector_params.omega_start,
-        data_collection_number=detector_params.run_number,  # type:ignore # the validator always makes this int
-        n_images=full_params.experiment_params.get_num_images(),
-        axis_range=full_params.experiment_params.image_width,
-        axis_end=(
-            full_params.experiment_params.omega_start
-            + full_params.experiment_params.rotation_angle
-        ),
-        kappa_start=full_params.experiment_params.chi_start,
-    )
-    (
-        info.xtal_snapshot1,
-        info.xtal_snapshot2,
-        info.xtal_snapshot3,
-    ) = get_xtal_snapshots(ispyb_params)
-    return info
-
-
-def construct_comment_for_rotation_scan() -> str:
-    return "Hyperion rotation scan"
